+++ ld_script.txt
@@ -557,6 +557,7 @@ SECTIONS {
         src/dungeon_menu_recruitment.o(.rodata);
         data/data_8106A4C.o(.rodata);
         src/code_8066D04.o(.rodata);
+        src/code_805D8C8.o(.rodata);
         data/data_8106A4C_2.o(.rodata);
         src/code_806CD90.o(.rodata);
         src/type_effectiveness.o(.data);
